pyproject.toml
[project]
name = "inventorymanagement"
version = "0.1.0"
description = "Inventory management system with Flask"
authors = [
    {name = "EmmyAnieDev",email = "emaduilzjr1@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "flask (>=3.1.0,<4.0.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "flask-migrate (>=4.1.0,<5.0.0)",
    "flask-jwt-extended (>=4.7.1,<5.0.0)",
    "flask-mail (>=0.10.0,<0.11.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "flask-marshmallow (>=1.3.0,<2.0.0)",
    "redis (>=6.0.0,<7.0.0)",
    "flask-restx (>=1.3.0,<2.0.0)",
    "flasgger (>=0.9.7.1,<0.10.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "marshmallow-sqlalchemy (>=1.4.2,<2.0.0)",
    "celery (>=5.5.2,<6.0.0)",
    "flower (>=2.0.1,<3.0.0)",
    "flask-cors (>=5.0.1,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
